Agents Screen fully done

+++ app/src/main/java/com/jasmeet/valorantapi/data/weaponData/WeaponStats.kt
@@ -9,10 +9,10 @@ data class WeaponStats(
     val equipTimeSeconds: Double,
     val feature: String,
     val fireMode: Any,
-    val fireRate: Int,
+    val fireRate: Double,
     val firstBulletAccuracy: Double,
     val magazineSize: Int,
-    val reloadTimeSeconds: Int,
+    val reloadTimeSeconds: Double,
     val runSpeedMultiplier: Double,
     val shotgunPelletCount: Int,
     val wallPenetration: String
